Update the project structure
