+++ components/IndexComponent/IndexComponent.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { ScrollView } from 'react-native';
-import { RoomsComponent } from './RoomsComponent/RoomsComponent';
-import { TemperatureAdjustComponent } from './TemperatureAdjust/TemperatureAdjust'
-import { RoomAccessoriesComponent } from './RoomAccessoriesComponent/RoomAccessoriesComponent';
+import { RoomsComponent } from './components/RoomsComponent/RoomsComponent';
+import { TemperatureAdjustComponent } from './components/TemperatureAdjust/TemperatureAdjust'
+import { RoomAccessoriesComponent } from './components/RoomAccessoriesComponent/RoomAccessoriesComponent';
 
 export class IndexComponent extends Component {
