+++ components/IndexComponent/components/RoomsComponent/RoomBox.js
