+++ components/IndexComponent/components/RoomsComponent/images/bed-active.png
