+++ components/IndexComponent/components/RoomsComponent/images/bed.png
