+++ components/IndexComponent/components/RoomsComponent/images/default-active.png
