+++ components/IndexComponent/components/RoomsComponent/images/default.png
