+++ components/IndexComponent/components/RoomsComponent/images/kitchen-active.png
