+++ components/IndexComponent/components/RoomsComponent/images/kitchen.png
