+++ components/IndexComponent/components/RoomsComponent/images/livingroom-active.png
