+++ components/mocks/RoomAccessoriesMock.js
